Credit/Debit Functionality and create Account request

+++ src/main/java/edu/asu/sbs/models/Transaction.java
@@ -37,7 +37,7 @@ public class Transaction implements Serializable {
     @NotNull
     @Column(nullable = false)
     @Min(1)
-    private Float transactionAmount;
+    private Double transactionAmount;
 
     @LastModifiedDate
     private Instant updatedTime;
